+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -14,7 +14,7 @@
   ([#5471](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5471))
 
 * Fixed the name of spans that have `http.request.method` attribute set to `_OTHER`.
-  Now, the span name have `HTTP` prefix, previously it was `_OTHER`.
+  The span name will be set as `HTTP {http.route}` as per the [specification](https://github.com/open-telemetry/semantic-conventions/blob/v1.24.0/docs/http/http-spans.md#name).
   ([#5484](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5484))
 
 ## 1.7.1